third_party/blink/web_tests/external/wpt/speculation-rules/prefetch/tentative/cookie-indices.https_include=changedWithRedirect2-expected.txt

This is a testharness.js-based test.
[FAIL] If the redirect needs to be rerequested but goes to the same place and that one doesn't vary, we actually can use the prefetched final response.
  assert_equals: expected (string) "hello" but got (undefined) undefined
Harness: the test ran to completion.
